// src/types/schemas/rightPicks.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type RightPicks = ValueLabelPair[];

export interface ValueLabelPair {
	value: string;
	label: string;
	[k: string]: unknown;
}